add default project

+++ src/modules/project.js
@@ -5,15 +5,15 @@ const projectFactory = (title, description, date, priority, completed, todos) =>
 
 const EXISTING_DATA = JSON.parse(localStorage.getItem('Projects')) || []
 
-const defaultProject = projectFactory('Default Project', 'Default project one', '08/02/2021', 'Medium', false );
 
 const defProject = () => {
-  for(let i = 0; i < EXISTING_DATA.length; i++){
-    while(EXISTING_DATA[i].title != 'Default Project'){
-      EXISTING_DATA.unshift(defaultProject)
-      localStorage.setItem('Projects', JSON.stringify(EXISTING_DATA))
-    }
+  let defaultProject = projectFactory('Default Project', 'Default project one', '08/02/2021', 'Medium', false, []);
+
+  if (!EXISTING_DATA.find(o => o.title === 'Default Project')){
+    EXISTING_DATA.push(defaultProject)
+    localStorage.setItem('Projects', JSON.stringify(EXISTING_DATA))
   }
+  
 }
 
  function deleteProject(index){
@@ -34,7 +34,6 @@ const defProject = () => {
     projectArr.push(project)
     EXISTING_DATA.push(project)
     localStorage.setItem('Projects', JSON.stringify(EXISTING_DATA))
-    uppdateProjectList()
 }
 
-export  { projectFactory, deleteProject, projectConstructor }
+export  { projectFactory, deleteProject, projectConstructor, defProject }
